+++ src/components/core/button.tsx
@@ -1,6 +1,12 @@
-import styled, { css } from 'styled-components'
+import styled from '@emotion/styled'
+import { css } from '@emotion/react'
 
-export const Button = styled.button`
+type ButtonProps = {
+  primary?: boolean
+  disabled?: boolean
+}
+
+export const Button = styled.button<ButtonProps>`
   display: flex;
   align-items: center;
   background-color: white;
@@ -25,6 +31,8 @@ export const Button = styled.button`
   ${props => props.primary && PrimaryButtonMixin}
   ${props => props.disabled && DisabledButtonMixin}
 `
+
+export const ButtonLink = Button.withComponent('a')
 
 const PrimaryButtonMixin = css`
   background-color: var(--bestofjsOrange);