+++ src/components/core/typography.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 /*
 Link to external websites, that open in a new browser tab