+++ src/components/home/progress-bar.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 export const ProgressBar = ({ progress }) => (
   <Container>