+++ src/components/project-details/npm-card/expandable-section.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 import { GoTriangleDown, GoTriangleRight } from 'react-icons/go'
 
 const TogglerLink = styled.a`