+++ src/components/tags/tag-list.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link, useHistory } from 'react-router-dom'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 import { useSelector } from 'containers/project-data-container'
 import { getProjectsByTag } from 'selectors'
@@ -28,7 +28,7 @@ export const CompactTagList = ({
   return (
     <TagList>
       {tags.map(tag => (
-        <ListRow key={tag.id} tag={tag}>
+        <ListRow key={tag.id}>
           <CompactListItem>
             <Link
               to={`/projects?tags=${tag.code}`}