+++ src/pages/timeline-page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 import { MainContent, PageTitle, Spinner } from 'components/core'
 import { allProjects } from 'selectors'